backend/requirements.txt
# Workspace version: 1.46.34
python-multipart==0.0.9
ipykernel==6.29.5
uvicorn[standard]==0.34.2
fastapi==0.115.12
databutton==0.38.34
openai
beautifulsoup4
requests
aiohttp
lxml
gspread==6.1.4
google-auth==2.35.0
pyairtable==2.3.3
psycopg2-binary==2.9.9
